Move data.ts to /utils, and fix intendation
